Raise exception when dividing by zero for ExampleApp

+++ ExampleGame/Assets/TestUI/TestUI.cs
@@ -27,7 +27,7 @@ public class TestUI : MonoBehaviour{
 		if(GUI.Button(GetControlRect(1), "Divide By Zero"))
 		{
 
-			int i = 0;
+			decimal i = 0;
 			i = 5 / i;
 		}
